CLI restructuring: - Rename BookCommands → BookCommandLine, move to cli.book package - Rename BookcaseCommands → BookcaseCommandLine, move to cli.bookcase package - Fix typo in prompt package: pompts → prompt - Rename test class to match: BookCommandsTest → BookCommandLineTest
Facade contract improvements:
- AuthorFacade.getOrCreateAuthorEntity() now returns Author (domain)
- ShelfFacade.getShelfEntityById() now returns Shelf (domain)
- Facades no longer leak infrastructure entities across domain boundaries

Cleanup:
- Remove Thread.sleep() calls from searchByAuthor() demo code
