+++ src/main/java/com/penrose/bibby/cli/prompt/api/PromptFacade.java
@@ -1,0 +1,4 @@
+package com.penrose.bibby.cli.prompt.api;
+
+public interface PromptFacade {
+}
