+++ src/main/java/com/penrose/bibby/library/author/api/AuthorFacade.java
@@ -4,5 +4,5 @@ import com.penrose.bibby.library.author.domain.Author;
 import com.penrose.bibby.library.author.infrastructure.entity.AuthorEntity;
 
 public interface AuthorFacade {
-    AuthorEntity getOrCreateAuthorEntity(Author author);
+    Author getOrCreateAuthorEntity(Author author);
 }
