+++ src/test/java/com/penrose/bibby/cli/BookCommandLineTest.java
@@ -3,7 +3,7 @@ package com.penrose.bibby.cli;
 import com.penrose.bibby.library.book.infrastructure.entity.BookEntity;
 import org.junit.jupiter.api.Test;
 
-class BookCommandsTest {
+class BookCommandLineTest {
 
     @Test
     public void searchByTitleTest(){
